+++ werobot/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.1.1'
+__version__ = '1.2.0'
 __author__ = 'whtsky'
 __license__ = 'MIT'