+++ werobot/client.py
@@ -55,6 +55,7 @@ class Client(object):
             **kwargs
         )
         r.raise_for_status()
+        r.encoding = "utf-8"
         json = r.json()
         if check_error(json):
             return json
@@ -557,7 +558,7 @@ class Client(object):
         return self.post(
             url="https://api.weixin.qq.com/cgi-bin/material/batchget_material",
             data={
-                "media": media_type,
+                "type": media_type,
                 "offset": offset,
                 "count": count
             }
@@ -708,7 +709,7 @@ class Client(object):
                 "user_list": [
                     {"openid": user_id,
                      "lang": lang} for user_id in user_id_list
-                    ]
+                ]
             }
         )