+++ werobot/utils.py
@@ -1,18 +1,26 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, unicode_literals
 
-import re
-import random
-import json
-import six
-import time
 import io
+import json
 import os
-
-from hashlib import sha1
+import random
+import re
+import string
+import time
 from functools import wraps
+from hashlib import sha1
+
+import six
+
+try:
+    from secrets import choice
+except ImportError:
+    from random import choice
 
 string_types = (six.string_types, six.text_type, six.binary_type)
+
+re_type = type(re.compile("regex_test"))
 
 
 def get_signature(token, timestamp, nonce, *args):
@@ -83,13 +91,8 @@ def generate_token(length=''):
         length = random.randint(3, 32)
     length = int(length)
     assert 3 <= length <= 32
-    token = []
-    letters = 'abcdefghijklmnopqrstuvwxyz' \
-              'ABCDEFGHIJKLMNOPQRSTUVWXYZ' \
-              '0123456789'
-    for _ in range(length):
-        token.append(random.choice(letters))
-    return ''.join(token)
+    letters = string.ascii_letters + string.digits
+    return ''.join(choice(letters) for _ in range(length))
 
 
 def json_loads(s):
@@ -140,3 +143,7 @@ def make_error_page(url):
             os.path.join(os.path.dirname(__file__), 'contrib/error.html'), 'r', encoding='utf-8'
     ) as error_page:
         return error_page.read().replace('{url}', url)
+
+
+def is_regex(value):
+    return isinstance(value, re_type)